moved ArrayTracker and BitMapTracker to their own files.
