+++ src/main/java/org/apache/commons/collections4/bloomfilter/Hasher.java
@@ -113,61 +113,5 @@ public interface Hasher {
             }
             return tracker.test(number) ? consumer.test(number) : true;
         }
-
-        /**
-         * An IndexTracker implementation that uses an array of integers to track whether or not a
-         * number has been seen.  Suitable for Shapes that have few hash functions.
-         * @since 4.5
-         */
-        static class ArrayTracker implements IntPredicate {
-            private int[] seen;
-            private int populated;
-
-            /**
-             * Constructs the tracker based on the shape.
-             * @param shape the shape to build the tracker for.
-             */
-            ArrayTracker(Shape shape) {
-                seen = new int[shape.getNumberOfHashFunctions()];
-            }
-
-            @Override
-            public boolean test(int number) {
-                if (number < 0) {
-                    throw new IndexOutOfBoundsException("number may not be less than zero. " + number);
-                }
-                for (int i = 0; i < populated; i++) {
-                    if (seen[i] == number) {
-                        return false;
-                    }
-                }
-                seen[populated++] = number;
-                return true;
-            }
-        }
-
-        /**
-         * An IndexTracker implementation that uses an array of bit maps to track whether or not a
-         * number has been seen.
-         * @since 4.5
-         */
-        static class BitMapTracker implements IntPredicate {
-            private long[] bits;
-
-            /**
-             * Constructs a bit map based tracker for the specified shape.
-             * @param shape The shape that is being generated.
-             */
-            BitMapTracker(Shape shape) {
-                bits = new long[BitMap.numberOfBitMaps(shape.getNumberOfBits())];
-            }
-
-            @Override
-            public boolean test(int number) {
-                boolean retval = !BitMap.contains(bits, number);
-                BitMap.set(bits, number);
-                return retval;
-            }
-        }
     }
 }
